Move drop_layer_dim_cap_data to other module to avoid circular import
