+++ imod/mf6/utilities/imod5_converter.py
@@ -6,9 +6,9 @@ import xarray as xr
 
 from imod.mf6.package import Package
 from imod.mf6.utilities.regrid import RegridderWeightsCache, _regrid_package_data
-from imod.msw.utilities.select import drop_layer_dim_cap_data
 from imod.typing import GridDataDict, Imod5DataDict
 from imod.typing.grid import full_like
+from imod.util.dims import drop_layer_dim_cap_data
 from imod.util.regrid_method_type import RegridMethodType
